% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp.R
\name{calc_echo_amp}
\alias{calc_echo_amp}
\title{Calculate echo amplitude from backscatter in an adcp object}
\usage{
calc_echo_amp(adcp)
}
\arguments{
\item{adcp}{}
}
\description{
Props to Ali Della Penna for help with this
}
