% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{center_longitude}
\alias{center_longitude}
\title{Center Longitude}
\usage{
center_longitude(elg)
}
\arguments{
\item{elg}{}
}
\description{
Experimental, A function that takes a vector of longitudes and centers them on the first longitude in the vector.
}
