% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{check_stations}
\alias{check_stations}
\title{Check Station}
\usage{
check_stations(data, summary, bottle = FALSE)
}
\arguments{
\item{bottle}{Logical, TRUE when processing a bottle datasheet. False all other times.}
}
\description{
Prechecks to make sure every station in the deployment sheet has a matching
station in the summary sheet. Sends a warning, not an error upon failure.
}
