% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{clean_bad_elg}
\alias{clean_bad_elg}
\title{Cut out bad lines of elg file based on bad nav gps signals}
\usage{
clean_bad_elg(filein, latstr = "[0-9]{4}\\\\.[0-9]{4}[NS]{1}",
  lonstr = "[0-9]{5}\\\\.[0-9]{4}[EW]{1}", fileout = NULL)
}
\arguments{
\item{filein}{.elg filepath to be read in}

\item{latstr}{regexp string to use to search for}

\item{fileout}{a specified output filepath.
If not specified, "_clean" is appended to name of input file and saved in the same directory.}
}
\description{
SEA event files contain output from a number of instruments
including GPS, flow-through, chirp, etc.
}
\examples{
clean_bad_elg()

}
