% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottle.R
\name{create_bottle}
\alias{create_bottle}
\title{Create bottle file sheet}
\usage{
create_bottle(
  summary_csv,
  ros_folder,
  datasheet_folder = NULL,
  csv_output = NULL
)
}
\arguments{
\item{datasheet_folder}{}
}
\value{

}
\description{
Create bottle file sheet
}
