% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{create_datasheet}
\alias{create_datasheet}
\title{Create complete data sheet from xls input}
\usage{
create_datasheet(
  data_input,
  summary_input = "output/csv/summary_datasheet.csv",
  data_type = "CTD",
  csv_folder = "output/csv",
  csv_filename = "datasheet.csv",
  odv_folder = "output/odv",
  odv_filename = "datasheet.txt",
  odv_export = TRUE,
  cruiseID = NULL,
  add_cruiseID = TRUE,
  add_deployment_type = TRUE,
  add_deployment_subfold = TRUE,
  preserve_col_names = FALSE,
  process_calc = FALSE,
  calc_folder = NULL,
  ...
)
}
\arguments{
\item{data_input}{File path for the .xls file with hand-recorded data values.
These are created by default in the "datasheets" folder and all have the
suffix "_input"}

\item{summary_input}{File path for the csv summary datasheet produced with
create_summary() (see details below). This defaults to the output/csv
folder in the standard SEA data project organization}

\item{data_type}{The data type code that will draw data from the summary
sheet (see details below)}

\item{csv_folder}{The directory path to output the csv file. Set to NULL for
no csv output.}

\item{csv_filename}{The csv filename to output the data}

\item{odv_folder}{The directory path to output the odv file. Set to NULL for
no odv output.}

\item{odv_filename}{The odv .txt filename to output the data}

\item{odv_export}{Logical defaults to TRUE. Will try to create an ODV
formatted datasheet. If the data type is not neuston, meter, bottle,
hydrocast, or CTD-- Seaprocess will use the format_gen_odv() function to
make an odv output.}

\item{cruiseID}{Optional string specifying cruise ID (i.e. "S301")}

\item{add_deployment_type}{logical to tell function whether to add the
default deployment type to the beginning of the filename that is being
created.}

\item{preserve_col_names}{logical. Function will default to renaming columns
so they are primarily alphanumeric and underscores. If you would like to
toggle this option "off", set to TRUE. The function will make your column
names lowercase but otherwise leave them unchanged. NOTE: This is only
recommended for datasheets that are not automatically created. Setting to
TRUE will cause errors for neuston, meter, bottle, secchi, ctd.}

\item{process_calc}{logical to tell the function whether to look for
calculation sheets to add to the datasheet. Likely this will mostly be used
with reef data. See ?seaprocess::read_calc_sheet_mb or the "Setup and EOC"
vignette for more information}

\item{calc_folder}{Defaults to NULL. If you wish to append calculation sheet
information to your datasheet, set this file path to a folder containing
all the calculation sheets you would like to use for that specific
datasheet. This should be autocreated within the 'datasheets' folder}

\item{...}{optional arguments to be sent to compile_bottle. Can be used to
process niskins on the wire. See the documentation for that function
(\code{?seaprocess::compile_bottle()}) for more information.}

\item{add_cruise_ID}{If cruiseID is set, logical to specify whether cruiseID
should be appended to beginning of filenames for csv and odv output}

\item{add_deployment_fold}{logical to tell function whether to add a new
directory to the end of the odv_folder directory path to keep .txt files
separate. Will create the new directory name depending on data_type}
}
\value{
If assigned to an object, the function will return the formatted
tibble that was exported to csv
}
\description{
This function reads in a hand-entered excel data sheet along with a station
summary csv file (created using \code{create_summary()}) and combines them to
produce well-formatted csv and odv file outputs.
}
\details{
data_input is the only argument that the function \emph{needs} to create an
output. The default values for summary_input and the csv and odv outputs are
set to work with the default directory configuration of the SEA cruise
project so shouldn't need to be set unless you are trying to produce
alternative outputs for specific custom cases or trials.

data_type can be either a stand-alone value entered into the deployment
column of the station summary sheet. Or it can take one of these special
cases:
\itemize{
\item "CTD" for ctd datasheets (default)
\item "neuston" for neuston datasheets
\item "bottle" for bottle datasheets
\item "meter" for meter net datasheets
\item "\if{html}{\out{<custom>}}" for custom datasheets you've created and given a unique code to in station summary
}

See "Introduction to SeaProcess" vignette for more information about
processing specific to neuston tows and meter nets.
}
