% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{create_gps_dttm}
\alias{create_gps_dttm}
\title{Create a GPS DateTime field}
\usage{
create_gps_dttm(gps_time, sys_dttm)
}
\arguments{
\item{gps_time}{Raw GPS time field}

\item{sys_dttm}{system datetime}
}
\value{

}
\description{
ELG files typically have a GPS time, but no GPS data.
This function takes the system datetime field already parsed and returns a well formated GPS datetime.
}
