% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cruise2ship}
\alias{cruise2ship}
\title{Cruise ID to ship name}
\usage{
cruise2ship(cruiseID)
}
\arguments{
\item{cruiseID}{string formatted as "C___" or S___"}
}
\description{
Takes a standard cruise ID and looks for an S or C at the beginning to create
a variable called "ship." Used to change processing based on other ship needs
since we use cruiseID in all functions.
}
