% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{ctd_to_tibble}
\alias{ctd_to_tibble}
\title{Convert a ctd profile from OCE class ctd to tibble}
\usage{
ctd_to_tibble(ctd_data, cruiseID = NULL, depth_vec = NULL, depth_step = 1)
}
\arguments{
\item{cruiseID}{}
}
\value{

}
\description{
Convert a ctd profile from OCE class ctd to tibble
}
