% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{eoc_process}
\alias{eoc_process}
\title{EOC Process}
\usage{
eoc_process(
  elg_folder = elg_folder,
  ctd_folder = ctd_folder,
  ros_folder = ros_folder,
  adcp_folder = adcp_folder,
  adcp_file_type = "LTA"
)
}
\arguments{
\item{elg_folder}{filepath to the folder containing elg files}

\item{ctd_folder}{filepath to the folder containing ctd cnv files}

\item{ros_folder}{filepath to the folder containing ros files}

\item{adcp_folder}{filepath to the folder containing adcp files}

\item{adcp_file_type}{type of ADCP data to copy. Defaults to "LTA" but can
accept "STA"}
}
\description{
Process EOC files
}
\details{
A simple function to copy relevant files to the "raw/" directory within the
cruise project. It's goal is to have a package of cruise data that is easy to
share with students and future scientists without requiring them to wade
through heaps of miscellaneous files. In it's current state will only copy
files needed to run the cruise processing scripts (.cnv, .ros .elg, .LTA (or
STA)).

Future improvements could include automatic outputs of cruise track,
distance traveled, deployment types etc.
}
