% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{fix_dateline}
\alias{fix_dateline}
\title{fix_dateline}
\usage{
fix_dateline(data)
}
\arguments{
\item{data}{}
}
\description{
Function to adjust mean longitude while crossing dateline.
}
\details{
This first creates a new variable "difflon" of the first lon - last lon in the
averaging window. This variable will shoot up to around 360 when the
antimeridian(dateline) is crossed. Next we filter out those time windows,
apply a correction, and then average *just* those specific lons. This allows
the rest of the averaging to happen through average_elg.
}
