% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{format_decimal}
\alias{format_decimal}
\title{Format to number of decimal places}
\usage{
format_decimal(df, col_name, digits)
}
\arguments{
\item{digits}{}
}
\description{
Rather than significant figures as format() provides in its digits = 2 option.
}
