% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odv.R
\name{format_elg_odv}
\alias{format_elg_odv}
\title{Format Hourly Data for ODV inport}
\usage{
format_elg_odv(data, odv_output = NULL, cruiseID = NULL)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Format Hourly Data for ODV inport
}
