% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odv.R
\name{format_gen_odv}
\alias{format_gen_odv}
\title{Format General ODV}
\usage{
format_gen_odv(data, file, cruiseID = NULL)
}
\arguments{
\item{data}{dataframe from csv output}

\item{file}{file path for odv output. Typcially will be
"output/odv/<cruiseID>_<yourdatasheet>.txt"}

\item{cruiseID}{optional to add cruiseID as a column in odv file.}
}
\description{
Function that takes a datasheet created by SeaProcess and using a large
lookup table, attempts to match common column names to well formatted odv
names. Any columns that do not match will be left as is. After the matching
occurs, it outputs a tab delimited file for import into ODV.
}
