% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odv.R
\name{format_meter_odv}
\alias{format_meter_odv}
\title{Format Meter Net Data for ODV import}
\usage{
format_meter_odv(data, file, cruiseID = NULL)
}
\arguments{
\item{data}{}
}
\description{
Format Meter Net Data for ODV import
}
