% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odv.R
\name{format_odv}
\alias{format_odv}
\title{Format}
\usage{
format_odv(
  data,
  folder,
  fields = c("hourly", "surfsamp", "neuston", "adcp", "ctd", "hydro"),
  import = FALSE,
  ...
)
}
\arguments{
\item{folder}{}
}
\value{

}
\description{
Format
}
