% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_elg}
\alias{get_elg}
\title{Get elg data from file or folder}
\usage{
get_elg(elg_input, ...)
}
\arguments{
\item{elg_input}{}
}
\description{
Takes an elg input and returns a data frame of all elg data fields before
they are averaged within the specified file or folder. Typically this will be
minute to minute data on SEA vessels.
}
