% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odv.R
\name{import_odv}
\alias{import_odv}
\title{Import a txt file to ODV}
\usage{
import_odv(odv_txt)
}
\arguments{
\item{odv_txt}{path to the odv txt file you want to import}
}
\value{

}
\description{
Import a txt file to ODV
}
