% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp.R
\name{interp_adcp}
\alias{interp_adcp}
\title{Interpolate an adcp field to different depths}
\usage{
interp_adcp(
  adcp,
  dvec,
  vars = c("u", "v", "Idb", "backscat", "quality", "percent")
)
}
\arguments{
\item{vars}{}
}
\value{

}
\description{
Interpolate an adcp field to different depths
}
