% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{interpolate_depth}
\alias{interpolate_depth}
\title{Interpolate a ctd tibble on a regular pressure grid to a regular dept grid}
\usage{
interpolate_depth(ctd_tibble, depth_vec = NULL)
}
\arguments{
\item{depth_vec}{}
}
\value{

}
\description{
Interpolate a ctd tibble on a regular pressure grid to a regular dept grid
}
