% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{message_read}
\alias{message_read}
\title{Prints message that states which slot is being read}
\usage{
message_read(field, ii)
}
\arguments{
\item{field}{The proper name of the field - gets passed to output string to better read}

\item{ii}{slot used for the parsing}
}
\description{
Called during parse_elg_* functions when field is being parsed
}
\examples{
message_read("Temperature",13)
}
