% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{parse_elg_field}
\alias{parse_elg_field}
\title{Generic parser for fields contained in araw charactor tibble}
\usage{
parse_elg_field(df, field, regex, parse_fun, ...)
}
\arguments{
\item{df}{character tibble}

\item{field}{name of field to parse (passed to warning and message functions for readability).}

\item{regex}{regex for finding the appropriate field. Can be vector of searches with earlier values prioritized.}

\item{parse_fun}{the parsing function to use in creating the new vector}

\item{...}{additional parameters passed to parse_fun (i.e. format = "" for parse_date)}
}
\description{
Data read in to character tibble by read_csv(...,col_types = cols(.default=col_character()))).
Takes parsing function as input so can differentiate between different column types.
Takes tibble and returns vector
}
\examples{
parse_elg_field()
}
