% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{parse_elg_gps}
\alias{parse_elg_gps}
\title{Parse GPS fields from ELG file}
\usage{
parse_elg_gps(df, source)
}
\arguments{
\item{df}{tibble character data frame input}

\item{source}{choose which GPS source to parse ("lab","nav")}
}
\description{
Takes in a raw character tibble data frame and parses the 6 potential GPS fields
Takes tibble, returns vector
}
\examples{
parse_gps()
}
