% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{parse_field}
\alias{parse_field}
\title{Generic field parser - can be used for all elg and excel parsing}
\usage{
parse_field(df, regex, parse_fun, ...)
}
\arguments{
\item{df}{input data frame containing column to format}

\item{regex}{regular expression to locate column name}

\item{parse_fun}{function to use in parsing data in function}

\item{...}{optional arguments to be passed to parse_fun}
}
\value{

}
\description{
Returns tibble with found column parsed and formatted
}
