% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{parse_lat}
\alias{parse_lat}
\title{Parse lat from elg file}
\usage{
parse_lat(latin)
}
\arguments{
\item{latin}{lat to process}
}
\description{
Wrapper function for parse_latlon() for latutude case
}
\examples{
parse_lat()

}
