% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{plot_ctd}
\alias{plot_ctd}
\title{Plot CTD}
\usage{
plot_ctd(ctd)
}
\arguments{
\item{ctd}{a formal 'ctd' object created by the oce package.}
}
\description{
Used to create ctd plot outputs in the RStudio Plots viewer. Ideally, marine
techs will be able to take a quick look at the most common parameters to
ensure profiles make sense. Currently setup to recognize temperature,
salinity, fluorescence and oxygen. If not all of those parameters are
present, it will default to plots that can be created using only temp,
conductivity, and pressure. Makes use of purrr functions to ensure that if an
error is thrown the ctd processing continues on.
}
\details{
Utilizes oce::plot() for all plotting. See the documentation for that
function for more information.
}
