% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot_limits}
\alias{plot_limits}
\title{Plot Limits}
\usage{
plot_limits(elg)
}
\arguments{
\item{elg}{}
}
\description{
Experimental, Takes an elg file and outputs a vector of the limits of the max
and min lats and lons with a buffer for plotting.
}
