% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_ctd}
\alias{process_ctd}
\title{Master processing function for ctd data}
\usage{
process_ctd(
  ctd_folder,
  cruiseID = NULL,
  csv_folder = "output/csv",
  csv_filename = "ctd.csv",
  odv_folder = "output/odv/ctd",
  odv_filename = "ctd.txt",
  add_cruiseID = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Master processing function for ctd data
}
