% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_elg}
\alias{process_elg}
\title{Master processing function for event data}
\usage{
process_elg(
  elg_folder,
  cruiseID = NULL,
  csv_folder = "output/csv",
  csv_filename = "elg.csv",
  odv_folder = "output/odv/elg",
  odv_filename = "elg.txt",
  add_cruiseID = TRUE,
  average_window = 60,
  ...
)
}
\arguments{
\item{elg_folder}{filepath to the folder containg elg files}

\item{cruiseID}{Optional string specifying cruise ID (i.e. "S301")}

\item{csv_folder}{The directory path to output the csv file. Set to NULL for
no csv output.}

\item{csv_filename}{The csv filename to output the data}

\item{odv_folder}{The directory path to output the odv file. Set to NULL for
no odv output.}

\item{odv_filename}{The odv .txt filename to output the data}

\item{average_window}{the averaging window in minutes for the exported file.
Set to NULL for no averaging. Default is 60 minutes.}

\item{add_cruise_ID}{If cruiseID is set, logical to specify whether cruiseID
should be appended to beginning of filenames for csv and odv output}
}
\value{

}
\description{
Function takes in a filepath to a folder containing the cruise elg files,
reads them in, combines them into one data object and then exports to a
comma-separated-file and an odv text file.
}
\details{
Note that the elg folder doesn not have to solely contain elg files, the
reading function will find all the elg files amoungst the other files
exported by SCS
}
