% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{process_wire}
\alias{process_wire}
\title{Process Wire Info}
\usage{
process_wire(data, raw_folder = NULL, cruiseID = NULL)
}
\arguments{
\item{data}{dataframe that has a dttm and dttm_out column}

\item{raw_folder}{folder path to where the raw tension files are stored. On
RCS this is typically on the datalogger computer with the name "LCI90-raw,
on CC this is the Hydrowinch Tension Files folder.}

\item{ship}{either "RCS" or "CC" used for dealing with the different file
names and tension logging methods between both boats.}
}
\description{
A function that works inside `create_summary()` to take a folder with raw
tension files, read it into R and find the tension peak and wire payout at
that peak for each deployment. It does this by aligning time in and time out
from the station summary sheet with the times in the tension files. This
means that on CC we must keep the LCI 90 time accurate and on RCS we must
keep the datalogger computer time accurate.
}
\details{
NOTE: on CC the function will try to read anything in the raw folder EXCEPT a
file with the name "archive files." On RCS it is looking for specifically the
string "LCI90-raw." If you are having trouble with this function try renaming
the files or reach out to the office.
}
