% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoc.R
\name{process_wire_log}
\alias{process_wire_log}
\title{Process Wire Log}
\usage{
process_wire_log(
  csv_folder = "output/csv",
  eoc_folder = "eoc",
  cruiseID = NULL,
  csv_filename = "wire_log.csv",
  add_cruiseID = TRUE
)
}
\arguments{
\item{csv_folder}{Where to find the datasheets to compile.}

\item{eoc_folder}{output for the final wire log.}

\item{cruiseID}{cruiseID . eg "S314"}

\item{csv_filename}{Optional filename for the log.}

\item{add_cruiseID}{Option to append cruiseID to the file name.}
}
\description{
Experimental. Cycles through all output csv sheets and compiles wire log data
for submission to UNOLS wire pool. Outputs to an EOC folder.
*Important* the function must see "max_wire_out_m" as a column name to
 include that deployment sheet in the final output.
}
