% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp.R
\name{read_adcp}
\alias{read_adcp}
\title{Read RDI ADCP ensemble file}
\usage{
read_adcp(adcp_file, calc_echo = FALSE)
}
\arguments{
\item{adcp_file}{The file name of the ensemble file}

\item{calc_echo}{Whether or not to calculate the echo intensity from the backscatter}
}
\value{

}
\description{
Read in a RDI Acoustic Current Dopler Current Profiler ensemble file with
extension .ENS, .ENR, .ENX, .STA, or .LTA.
}
\details{
This makes use of the read.adp() function from the oce package which decodes
the binary format of a RDI ADCP and creates an oce adcp object. This is then
converted to a more readily used format, a list of available fields.
}
