% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp.R
\name{read_adcp_fold}
\alias{read_adcp_fold}
\title{Read all ADCP Ensemble files in a folder and combine}
\usage{
read_adcp_fold(
  adcp_fold,
  file_type = "(.LTA)|(.STA)|(.ENS)|(.ENX)",
  file_select = NULL,
  calc_echo = FALSE,
  combine = FALSE,
  combine_na = TRUE,
  sort_dttm = FALSE,
  ...
)
}
\arguments{
\item{adcp_fold}{}
}
\value{

}
\description{
Read all ADCP Ensemble files in a folder and combine
}
