% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcsheet.R
\name{read_calc_fold_mb}
\alias{read_calc_fold_mb}
\title{#Read a folder of calculation sheets, and append the data to the bottle
summary}
\usage{
read_calc_fold_mb(calc_folder, output)
}
\arguments{
\item{calc_folder}{#the calculation folder to be read. chlorophyll, PO4, NO3,
pH. All excel files must be formatted correctly with an 'output' sheet.}
}
\description{
This function reads the [output] sheet of every excel file in a folder and
appends the new values to the bottle summary object [bottle_sum].
}
\details{
This function, and the [read_calc_fold_mb()] function combine the
 [output] sheet from any calculation sheet (from path ['calc_file']) and the
 [bottle_input] datasheet. The function will only fill values where there is
 an "NA" so any numerical values entered in bottle_input will not be
 overwritten. This function will read every excel file in the specified
 folder [calc_folder] and append the [output] sheet to the bottle summary
 csv. Common errors arise from spelling typos in the header data of the many
 calculation excel files, or from unexpected values in the columns
 (characters vs. numbers etc.). Individual files may be run with the
 [read_calc_file()] function for troubleshooting.
}
