% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcsheet.R
\name{read_calc_sheet_mb}
\alias{read_calc_sheet_mb}
\title{Append calculation sheet data to bottle summary}
\usage{
read_calc_sheet_mb(calc_file, output)
}
\arguments{
\item{calc_file}{#the calculation file to be read, chlorophyll, PO4, NO3, pH.
This file must be formatted correctly with an 'output' sheet containing only
station number, bottle number, and paramter name (chla, no3 etc).}

\item{output}{#the datasheet to output calc sheet data to. This will typically
be the bottle_input sheet.}
}
\description{
This function appends hand-entered and calculated data from calculation sheets
to the bottle summary csv. It takes the 'output' sheet on the calculation
excel file with whatever header value is used and matches it with station
number and bottle number.
}
\details{
This function, and the [read_calc_fold_mb()] function combine the
[output] sheet from any calculation sheet (from path ['calc_file']) and the
[bottle_input] datasheet. The function will only fill values where there is an
"NA" so any numerical values entered in bottle_input will not be overwritten.
}
