% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcsheet.R
\name{read_calc_sheet_net}
\alias{read_calc_sheet_net}
\title{Read Net Calculation Sheet}
\usage{
read_calc_sheet_net(calc_file, output)
}
\arguments{
\item{output}{}
}
\description{
Not in use. Similar to our other calculation sheet function but does not
require or look for a column labelled "bottle." Instead your output sheet
would have just 'station' and 'parameter' to worry about.
}
