% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{read_cnv_latlon}
\alias{read_cnv_latlon}
\title{Read lat and lon from CNV header}
\usage{
read_cnv_latlon(cnv_file)
}
\arguments{
\item{cnv_file}{cnv file to read}
}
\description{
Read lat and lon from CNV header
}
