% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{read_ctd_fold}
\alias{read_ctd_fold}
\title{Read all CNV files in a folder}
\usage{
read_ctd_fold(fold, check_vars = TRUE, cruiseID = NULL, depth_vec = NULL, ...)
}
\arguments{
\item{fold}{folder to read from}

\item{...}{}
}
\value{

}
\description{
Read all CNV files in a folder
}
