% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{read_elg}
\alias{read_elg}
\title{Read in SEA data from an ELG event file and return a well formatted output}
\usage{
read_elg(
  filein,
  forceGPS = NULL,
  preCheck = TRUE,
  skip = 0,
  csv_output = NULL,
  keep = c("dttm", "lon", "lat", "temp", "sal", "fluor", "cdom", "xmiss", "wind_sp",
    "wind_dir", "heading", "pitch", "roll", "bot_depth", "filename")
)
}
\arguments{
\item{filein}{Filepath of the elg file to be read in}

\item{forceGPS}{option to force the longitude, latitude and time to come from one
or other of the GPS feeds.}

\item{preCheck}{option to do an initial check of lines to remove any standard issues}

\item{skip}{number of lines to skip below header}

\item{csv_output}{optional path to export the processed data to a csv file}
}
\description{
SEA event files contain output from a number of instruments
including GPS, flow-through, chirp, etc.
}
\examples{
read_elg()

}
