% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{read_elg_fold}
\alias{read_elg_fold}
\title{Read multiple ELG files from a folder}
\usage{
read_elg_fold(root_folder, sort_elg = TRUE, ...)
}
\arguments{
\item{root_folder}{}
}
\description{
Read multiple ELG files from a folder
}
