% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{update_elg}
\alias{update_elg}
\title{Update ELG data}
\usage{
update_elg(df, filein, preCheck = T)
}
\arguments{
\item{df}{current tibble of ELG data}

\item{filein}{location of the *.elg file to update from}
}
\description{
Add to an existing elg tibble from an SEA Event File that is in the process of being recorded.
}
