% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{warning_empty}
\alias{warning_empty}
\title{Produces warning for an empty field}
\usage{
warning_empty(field)
}
\arguments{
\item{field}{The proper name of the field - gets passed to output string to better read}
}
\description{
Called during parse_elg_* functions if no match to the input regex is found
}
\examples{
warning_empty("Temperature")
}
