% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{warning_mult}
\alias{warning_mult}
\title{Produces warning for multiple name matches}
\usage{
warning_mult(field, ii)
}
\arguments{
\item{field}{The proper name of the field - gets passed to output string to better read}

\item{ii}{matched field names}
}
\description{
Called during parse_elg_* functions if multiple matches to the input regex are found
NB: by default the first value is selected.
}
