% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odv.R
\name{write_odv}
\alias{write_odv}
\title{Write formatted data to odv txt file}
\usage{
write_odv(odv_out, odv_output = NULL)
}
\arguments{
\item{odv_output}{}
}
\value{

}
\description{
Write formatted data to odv txt file
}
