% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{wswd_to_uv}
\alias{wswd_to_uv}
\title{Convert wind speed and wind direction to u and v}
\usage{
wswd_to_uv(ws, wd)
}
\arguments{
\item{ws}{input wind speed vector}

\item{wd}{input wind direction vector}
}
\value{

}
\description{
Convert wind speed and wind direction to u and v
}
