% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{compile_bottle}
\alias{compile_bottle}
\title{Create bottle file sheet}
\usage{
compile_bottle(data, ros_input)
}
\arguments{
\item{datasheet_folder}{}
}
\value{

}
\description{
Create bottle file sheet
}
