% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{create_datasheet}
\alias{create_datasheet}
\title{Create complete data sheet}
\usage{
create_datasheet(
  data_input,
  summary_input = "output/csv/summary_datasheet.csv",
  data_type = "CTD",
  csv_folder = "output/csv",
  csv_filename = "datasheet.csv",
  odv_folder = "output/odv",
  odv_filename = "datasheet.txt",
  cruiseID = NULL,
  add_cruiseID = TRUE,
  add_deployment_type = TRUE,
  add_deployment_subfold = TRUE,
  ...
)
}
\arguments{
\item{data_input}{Filepath for the xls file with hand-recorded data values}

\item{summary_input}{Filepath for the csv summary datasheet produced with
create_summary() (see details below)}

\item{data_type}{The data type code that will draw data from the summary
sheet (see details below)}

\item{csv_folder}{The directory path to output the csv file. Set to NULL for
no csv output.}

\item{csv_filename}{The csv filename to output the data}

\item{odv_folder}{The directory path to output the odv file. Set to NULL for
no odv output.}

\item{odv_filename}{The odv .txt filename to output the data}

\item{cruiseID}{Optional string specifying cruise ID (i.e. "S301")}

\item{add_deployment_type}{logical to tell function whether to add a new
directory to the end of the odv_folder directory path to keep .txt files
separate. Will create the new directory name depending on data_type}

\item{add_cruise_ID}{If cruiseID is set, logical to specify whether cruiseID
should be appended to beginning of filenames for csv and odv output}
}
\value{

}
\description{
This function reads in a hand-entered excel data sheet along with a station
summary csv and combines them to produce well-formatted csv and odv file
outputs.
}
\details{
data_input is the only argument that the function *needs* to create an output. The default values for summary_input and the csv and odv outputs are set to work with the default directory configuration of the SEA cruise project so shouldn't need to be set unless you are trying to produce alternative outputs for specific custom cases.
}
