% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheets.R
\name{create_datasheet}
\alias{create_datasheet}
\title{Create complete data sheet}
\usage{
create_datasheet(data_input, summary, data_type = "ctd",
  csv_output = NULL)
}
\arguments{
\item{summary}{}
}
\description{
Create complete data sheet
}
