% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{create_datasheet}
\alias{create_datasheet}
\title{Create complete data sheet from xls input}
\usage{
create_datasheet(
  data_input,
  summary_input = "output/csv/summary_datasheet.csv",
  data_type = "CTD",
  csv_folder = "output/csv",
  csv_filename = "datasheet.csv",
  odv_folder = "output/odv",
  odv_filename = "datasheet.txt",
  cruiseID = NULL,
  add_cruiseID = TRUE,
  add_deployment_type = TRUE,
  add_deployment_subfold = TRUE,
  ...
)
}
\arguments{
\item{data_input}{File path for the .xls file with hand-recorded data values. These are created by default in the "datasheets" folder and all have the suffix "_input"}

\item{summary_input}{File path for the csv summary datasheet produced with
create_summary() (see details below). This defaults to the output/csv folder in the standard SEA data project organization}

\item{data_type}{The data type code that will draw data from the summary
sheet (see details below)}

\item{csv_folder}{The directory path to output the csv file. Set to NULL for
no csv output.}

\item{csv_filename}{The csv filename to output the data}

\item{odv_folder}{The directory path to output the odv file. Set to NULL for
no odv output.}

\item{odv_filename}{The odv .txt filename to output the data}

\item{cruiseID}{Optional string specifying cruise ID (i.e. "S301")}

\item{add_deployment_type}{logical to tell function whether to add the default deployment type to the beginning of the filename that is being created.}

\item{...}{option arguments to be sent to compile_bottle. Initially, this is
just ros_input}

\item{add_cruise_ID}{If cruiseID is set, logical to specify whether cruiseID
should be appended to beginning of filenames for csv and odv output}

\item{add_deployment_fold}{logical to tell function whether to add a new
directory to the end of the odv_folder directory path to keep .txt files
separate. Will create the new directory name depending on data_type}
}
\value{
If assigned to an object, the function will return the formatted
tibble that was exported to csv
}
\description{
This function reads in a hand-entered excel data sheet along with a station
summary csv file (created using \code{create_summary()}) and combines them to produce well-formatted csv and odv file
outputs.
}
\details{
data_input is the only argument that the function \emph{needs} to create an
output. The default values for summary_input and the csv and odv outputs are
set to work with the default directory configuration of the SEA cruise
project so shouldn't need to be set unless you are trying to produce
alternative outputs for specific custom cases or trials.

data_type can be either a stand-alone value entered into the deployment
column of the station summary sheet. Or it can take one of these special
cases:
\itemize{
\item "CTD" for ctd datasheets (default)
\item "neuston" for neuston datasheets
\item "bottle" for bottle datasheets
\item "meter" for meter net datasheets
\item "<custom>" for custom datasheets you've created and given a unique code to in station summary
}
}
