% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{create_datasheet}
\alias{create_datasheet}
\title{Create complete data sheet}
\usage{
create_datasheet(
  data_input,
  summary_input = "output/csv/summary_datasheet.csv",
  data_type = "CTD",
  csv_folder = "output/csv",
  csv_filename = "datasheet.csv",
  cruiseID = NULL,
  add_cruiseID = TRUE,
  add_deployment_type = TRUE,
  ...
)
}
\arguments{
\item{summary}{This function reads in a hand-entered excel sheet along with the summary CSV and produces a well formated CSV with all the relevent data
and meta data.}
}
\value{

}
\description{
Create complete data sheet
}
