% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{create_summary}
\alias{create_summary}
\title{Create Station Summary Sheet}
\usage{
create_summary(summary_input, elg_input, csv_output = NULL)
}
\arguments{
\item{summary_input}{The input datasheet that includes the relevent station and deployment metadata}

\item{elg}{The cruise elg file or folder or files for extracting continuous data from}
}
\description{
This function combines hand-entered station metadata with electronically recorded instrument data to output a well formatted station summary sheet.
}
\examples{
# Create a summary from S285 data input and elgs
summary_input <- system.file("extdata", "S285_station.xlsx", package="seaprocess")
elg_input <- system.file("extdata", "S285_elg", package="seaprocess")
summary <- create_summary(summary_input, elg_input)
}
