% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{create_summary}
\alias{create_summary}
\title{Create Station Summary Sheet}
\usage{
create_summary(
  summary_input,
  elg_input,
  csv_output = NULL,
  force_stations = FALSE,
  ...
)
}
\arguments{
\item{summary_input}{The input datasheet that includes the relevent station and deployment metadata}

\item{csv_output}{The desired output location for the well-formatted csv file}

\item{elg}{The cruise elg file (or folder of files) for extracting continuous data from}
}
\value{
A tibble containing the combined data frames. If csv_output is set to a valid output path then a formatted csv file is output also.
}
\description{
This function combines hand-entered station metadata with electronically recorded location and environmental data to output a well formatted station summary sheet.
}
\details{
During deployments at SEA, we maintain paper data sheets which recorded station metadata and data from that deployment. These are a vital component for our data accuracy and redundancy.

In creating electronic datasheets for deployments it is desirable to combine hand-entered station metadata (station number, station type, zone description, etc.) with electronically recorded data (location, surface conditions, etc.) to provide an accurate record of the deployment without the need to re-enter hand-recorded values of the electronic data.

To do this, \code{\link[=create_summary]{create_summary()}} takes in an excel sheet with the bare minimum of hand-entered data:
\itemize{
\item Station number
\item Deployment type
\item Deployment date/time in (and out)
\item Zone description (time zone)
}

\code{\link[=create_summary]{create_summary()}} combines this data with the electronically recorded "event" data which records (amongst other things):
\itemize{
\item Time
\item Location
\item Surface Temperature
\item Surface Salinity
\item etc.
}

The event data is typically stored in a file with an extension elg. \code{\link[=read_elg]{read_elg()}} deals with reading this data in and formatting it properly.

Once the two data frames are read in to R, they are combined using the UTC time that exists in both data frames.
}
\examples{
# Create a summary from S285 data input and elgs
summary_input <- system.file("extdata", "S285_station.xlsx", package="seaprocess")
elg_input <- system.file("extdata", "S285_elg", package="seaprocess")
summary <- create_summary(summary_input, elg_input)
}
