% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odv.R
\name{format_odv}
\alias{format_odv}
\title{Wrapper function to format any odv file}
\usage{
format_odv(data, odv_output, data_type = "CTD", cruiseID = NULL)
}
\arguments{
\item{cruiseID}{}
}
\value{

}
\description{
Wrapper function to format any odv file
}
