% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elg.R
\name{parse_field}
\alias{parse_field}
\title{Generic field parser - can be used for all elg and excel parsing}
\usage{
parse_field(df, regex, parse_fun, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Returns tibble with found column parsed and formatted
}
