% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_adcp}
\alias{process_adcp}
\title{Master processing function for ADCP data}
\usage{
process_adcp(
  adcp_folder,
  cruiseID = NULL,
  csv_folder = "output/csv",
  csv_filename = "adcp.csv",
  odv_folder = "output/odv/adcp",
  odv_filename = "adcp.txt",
  add_cruiseID = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Master processing function for ADCP data
}
