% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_elg}
\alias{process_elg}
\title{Title}
\usage{
process_elg(
  elg_folder,
  output_csv = NULL,
  output_odv = NULL,
  average_window = 60,
  cruiseID = NULL,
  ...
)
}
\arguments{
\item{output_file}{}
}
\value{

}
\description{
Title
}
