% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_elg}
\alias{process_elg}
\title{Master processing function for event data}
\usage{
process_elg(
  elg_folder,
  cruiseID = NULL,
  csv_folder = "output/csv",
  csv_filename = "elg.csv",
  odv_folder = "output/odv/elg",
  odv_filename = "elg.txt",
  add_cruiseID = TRUE,
  average_window = 60,
  ...
)
}
\arguments{
\item{output_file}{}
}
\value{

}
\description{
Master processing function for event data
}
